monitor tokio runtime and handle failures
We use an erlang process to monitor the runtime, so that its failure
can be handled by a supervisor
