+++ rebar.config
@@ -8,3 +8,5 @@
 {profiles, [{test, [{deps, [ {jsx, "3.1.0"}
                            , {erlexec, "1.18.11"}
                            ]}]}]}.
+
+{shell, [{apps, [erqwest]}]}.
